# docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/analytics/requestor-summary": {
      "get": {
        "operationId": "AnalyticsController_requestorSummary",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "Acting role for RBAC. Allowed values: Requestor, Dept Head, Registrar, Staff, System Admin.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Staff"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Acting user id used for department scoping and own-record views. Example: U100.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "U100"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Requestor summary returned successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestorSummaryResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error or invalid workflow transition.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Role or department scope does not allow this action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Requested entity was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "role-header": []
          },
          {
            "user-id-header": []
          }
        ],
        "summary": "Get requestor dashboard summary",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/analytics/department-summary": {
      "get": {
        "operationId": "AnalyticsController_departmentSummary",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "Acting role for RBAC. Allowed values: Requestor, Dept Head, Registrar, Staff, System Admin.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Staff"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Acting user id used for department scoping and own-record views. Example: U100.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "U100"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Department summary returned successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentSummaryResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error or invalid workflow transition.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Role or department scope does not allow this action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Requested entity was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "role-header": []
          },
          {
            "user-id-header": []
          }
        ],
        "summary": "Get department-head dashboard summary",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/analytics/registrar-summary": {
      "get": {
        "operationId": "AnalyticsController_registrarSummary",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "Acting role for RBAC. Allowed values: Requestor, Dept Head, Registrar, Staff, System Admin.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Staff"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Acting user id used for department scoping and own-record views. Example: U100.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "U100"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Registrar summary returned successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrarSummaryResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error or invalid workflow transition.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Role or department scope does not allow this action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Requested entity was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "role-header": []
          },
          {
            "user-id-header": []
          }
        ],
        "summary": "Get registrar dashboard summary",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/analytics/stock": {
      "get": {
        "operationId": "AnalyticsController_stock",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "Acting role for RBAC. Allowed values: Requestor, Dept Head, Registrar, Staff, System Admin.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Staff"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Acting user id used for department scoping and own-record views. Example: U100.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "U100"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stock monitoring rows returned successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StockRowResponseDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error or invalid workflow transition.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Role or department scope does not allow this action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Requested entity was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "role-header": []
          },
          {
            "user-id-header": []
          }
        ],
        "summary": "Get stock threshold rows visible to the acting role",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/analytics/stock-thresholds/{id}/{level}": {
      "patch": {
        "operationId": "AnalyticsController_updateStockThreshold",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "Acting role for RBAC. Allowed values: Requestor, Dept Head, Registrar, Staff, System Admin.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Staff"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Acting user id used for department scoping and own-record views. Example: U100.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "U100"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": "STK-1001",
              "type": "string"
            }
          },
          {
            "name": "level",
            "required": true,
            "in": "path",
            "schema": {
              "example": 10,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stock threshold updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockRowResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error or invalid workflow transition.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Role or department scope does not allow this action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Requested entity was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "role-header": []
          },
          {
            "user-id-header": []
          }
        ],
        "summary": "Update a stock threshold level",
        "tags": [
          "analytics"
        ]
      }
    },
...
  },
  "info": {
    "title": "ResourceX Demo Backend",
    "description": "Demo-exact NestJS backend for the ResourceX frontend. All authorization is role-based through request headers. Use x-user-role on every protected route and x-user-id when actor-specific scoping is required.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api",
      "description": "API base path"
    }
  ],
  "components": {
    "securitySchemes": {
      "role-header": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-role",
        "description": "Required RBAC header. Example values: Requestor, Dept Head, Registrar, Staff, System Admin."
      },
      "user-id-header": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-id",
        "description": "Optional acting user id header for own-profile and department-scoped behavior. Example: U100."
      }
    },
    "schemas": {
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "message": {
            "type": "object",
            "example": "Selected resources must match request quantity."
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
...
      "ProcurementResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "PROC-819"
          },
          "resourceType": {
            "type": "string",
            "example": "Server Blades v2"
          },
          "item": {
            "type": "string",
            "example": "Server Blades v2"
          },
          "quantity": {
            "type": "number",
            "example": 10
          },
          "department": {
            "type": "string",
            "example": "IT Services"
          },
          "requestedBy": {
            "type": "string",
            "example": "pradhyum"
          },
          "requester": {
            "type": "string",
            "example": "pradhyum"
          },
          "requestedById": {
            "type": "string",
            "example": "U2"
          },
          "status": {
            "type": "string",
            "example": "Pending"
          },
          "date": {
            "type": "string",
            "example": "Oct 26, 2023"
          },
          "justification": {
            "type": "string",
            "example": "Capacity increase required."
          },
          "vendor": {
            "type": "string",
            "example": "Dell"
          },
          "invoice": {
            "type": "string",
            "example": "INV-2026-1001"
          }
        },
        "required": [
          "id",
          "resourceType",
          "quantity",
          "department",
          "status",
          "date",
          "justification"
        ]
      },
...
      "RequestorSummaryResponseDto": {
        "type": "object",
        "properties": {
          "totalRequests": {
            "type": "number",
            "example": 4
          },
          "pendingRequests": {
            "type": "number",
            "example": 2
          },
          "approvedRequests": {
            "type": "number",
            "example": 1
          },
          "allocatedRequests": {
            "type": "number",
            "example": 1
          },
          "resources": {
            "type": "number",
            "example": 2
          },
          "maintenance": {
            "type": "number",
            "example": 1
          }
        },
        "required": [
          "totalRequests",
          "pendingRequests",
          "approvedRequests",
          "allocatedRequests",
          "resources",
          "maintenance"
        ]
      },
      "DepartmentSummaryResponseDto": {
        "type": "object",
        "properties": {
          "totalResources": {
            "type": "number",
            "example": 12
          },
          "available": {
            "type": "number",
            "example": 7
          },
          "allocated": {
            "type": "number",
            "example": 3
          },
          "maintenance": {
            "type": "number",
            "example": 1
          },
          "scrap": {
            "type": "number",
            "example": 1
          },
          "pendingRequests": {
            "type": "number",
            "example": 2
          },
          "procurements": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProcurementResponseDto"
            }
          }
        },
        "required": [
          "totalResources",
          "available",
          "allocated",
          "maintenance",
          "scrap",
          "pendingRequests",
          "procurements"
        ]
      },
      "RegistrarSummaryResponseDto": {
        "type": "object",
        "properties": {
          "totalAssets": {
            "type": "number",
            "example": 14
          },
          "pendingRequests": {
            "type": "number",
            "example": 4
          },
          "pendingProcurements": {
            "type": "number",
            "example": 3
          },
          "approvedToday": {
            "type": "number",
            "example": 5
          }
        },
        "required": [
          "totalAssets",
          "pendingRequests",
          "pendingProcurements",
          "approvedToday"
        ]
      },
      "StockRowResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "TH-1"
          },
          "department": {
            "type": "string",
            "example": "IT Services"
          },
          "resourceType": {
            "type": "string",
            "example": "Laptop"
          },
          "thresholdLevel": {
            "type": "number",
            "example": 15
          },
          "currentQuantity": {
            "type": "number",
            "example": 18
          },
          "status": {
            "type": "string",
            "example": "Safe"
          }
        },
        "required": [
          "id",
          "department",
          "resourceType",
          "thresholdLevel"
        ]
      },
...
    }
  }
}